Separates dev and production npm modules

+++ frontend.sh
@@ -11,6 +11,30 @@ set -e
 
 # Initialize project dependency directories.
 init(){
+  # Set default command-line environment flag, if user didn't supply one.
+  if [ -z "$1" ]; then
+    cli_flag='development'
+  else
+    cli_flag=$1
+  fi
+
+  # Warn if unsupported command-line flag was used.
+  if [ "$cli_flag" != "development" ] &&
+     [ "$cli_flag" != "test" ] &&
+     [ "$cli_flag" != "production" ]; then
+    echo "\033[33;33mWARNING: '$cli_flag' flag not found, reverting to development environment.\033[0m"
+    cli_flag='development'
+  fi
+
+  # Notify of environment that user is in.
+  if [ "$cli_flag" = "development" ]; then
+    echo 'Frontend:\033[33;33m development environment\033[0m.'
+  elif [ "$cli_flag" = "test" ]; then
+    echo 'Frontend:\033[33;1m test environment\033[0m.'
+  elif [ "$cli_flag" = "production" ]; then
+    echo 'Frontend:\033[32;1m production environment\033[0m.'
+  fi
+
   ENVVAR_SAMPLE=.env_SAMPLE
   ENVVAR=.env
   NODE_DIR=node_modules
@@ -52,25 +76,32 @@ install(){
 
   # Copy globally-installed packages.
   # Protractor - JavaScript acceptance testing.
-  if [ $(is_installed protractor) = 0 ]; then
-    echo 'Installing Protractor dependencies locally...'
-    npm install protractor
-    ./$NODE_DIR/protractor/bin/webdriver-manager update
-  else
-    echo 'Global Protractor installed. Copying global install locally...'
-    protractor_symlink=$(command -v protractor)
-    protractor_binary=$(readlink $protractor_symlink)
-    protractor_full_path=$(dirname $protractor_symlink)/$(dirname $protractor_binary)/../../protractor
-    if [ ! -d $protractor_full_path/selenium ]; then
-      echo '\033[33;31m ERROR: Please run `webdriver-manager update` and try again!'
-      exit
+  if [ "$cli_flag" = "development" ] ||
+     [ "$cli_flag" = "test" ]; then
+    # Copy globally-installed packages.
+    # Protractor - JavaScript acceptance testing.
+    if [ $(is_installed protractor) = 0 ]; then
+      echo 'Installing Protractor dependencies locally...'
+      npm install protractor
+      ./$NODE_DIR/protractor/bin/webdriver-manager update
+    else
+      echo 'Global Protractor installed. Copying global install locally...'
+      protractor_symlink=$(command -v protractor)
+      protractor_binary=$(readlink $protractor_symlink)
+      protractor_full_path=$(dirname $protractor_symlink)/$(dirname $protractor_binary)/../../protractor
+      if [ ! -d $protractor_full_path/selenium ]; then
+        echo '\033[33;31mERROR: Please run `webdriver-manager update` and try again!\033[0m'
+        exit
+      fi
+      mkdir -p ./$NODE_DIR/protractor
+      cp -r $protractor_full_path ./$NODE_DIR/
+      rm -rf ./$NODE_DIR/protractor/node_modules/
     fi
-    mkdir -p ./$NODE_DIR/protractor
-    cp -r $protractor_full_path ./$NODE_DIR/
-    rm -rf ./$NODE_DIR/protractor/node_modules/
-  fi
 
-  npm install
+    npm install -d
+  else
+    npm install --production
+  fi
   bower install --config.interactive=false
 }
 
@@ -93,7 +124,7 @@ is_installed(){
   echo "$return_"
 }
 
-init
+init "$1"
 clean
 install
 build
